+++ framework/Steps/PostProcess.py
@@ -1,0 +1,32 @@
+# Copyright 2017 Battelle Energy Alliance, LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+  PostProcess module
+  This module contains the step that is aimed to PostProcess the
+  results of a RAVEN (or not) analysis.
+  Created on May 6, 2021
+  @author: alfoa
+  supercedes Steps.py from alfoa (2/16/2013)
+"""
+#External Modules------------------------------------------------------------------------------------
+#External Modules End--------------------------------------------------------------------------------
+
+#Internal Modules------------------------------------------------------------------------------------
+from .SingleRun import SingleRun
+#Internal Modules End--------------------------------------------------------------------------------
+
+class PostProcess(SingleRun):
+  """
+    This is an alternate name for SingleRun
+  """
